refactor(validators): 도메인 기반 구조로 리팩토링
- domains/todo/ 폴더 구조로 스키마 분리
  - todo.schema.ts: 엔티티 스키마 (todoSchema, Todo 타입)
  - todo.dto.ts: Request DTO 스키마 (create, update)
- API common과 중복되는 response.ts, error.ts 삭제
- utils/ 폴더 추가 (향후 확장용)
- 기존 export 호환성 유지
