+++ packages/validators/src/domains/todo/index.ts
@@ -1,0 +1,2 @@
+export * from './todo.dto';
+export * from './todo.schema';
