+++ packages/validators/src/utils/index.ts
@@ -1,0 +1,3 @@
+/**
+ * 유틸리티 함수 (향후 확장용)
+ */
